refactor: segment aws files
