+++ AWS/DynamoDB/cards.py
@@ -1,5 +1,5 @@
 from .user import get_user
-from ..db import users_table, access_log_table, thread_logger
+from ..db import users_table, thread_logger
 
 def register_card_to_user(user_id: str, card_id: str):
     user = get_user(user_id)
