+++ AWS/DynamoDB/user.py
@@ -1,14 +1,5 @@
-import datetime
-import random
-import time
-from typing import Optional
-
-from ..db import users_table, access_log_table, thread_logger
-
-def generate_unique_id():
-    timestamp = int(time.time() * 1000)
-    random_number = random.randint(1000, 9999)
-    return f"{timestamp}{random_number}"
+from Util import generate_unique_id
+from ..db import users_table, thread_logger
 
 def register_user(name: str):
     user_id = generate_unique_id()
@@ -31,6 +22,12 @@ def get_user(user_id: str):
     response = users_table.get_item(Key={"UserID": user_id})
     thread_logger.info(f"Get_User RESPONSE ->>>> {response.get('Item')}")
     return response.get('Item')
+
+def get_all_users():
+    thread_logger.info("Attempting to retrieve all users.")
+    response = users_table.scan()
+    thread_logger.info(f"Get_All_Users RESPONSE ->>>> {response.get('Items')}")
+    return response.get('Items')
 
 def edit_user(user_id: str, name: str = None, card_id: str = None, last_scanned: str = None):
     existing_user = get_user(user_id)
@@ -68,32 +65,3 @@ def edit_user(user_id: str, name: str = None, card_id: str = None, last_scanned:
     updated_item = response.get("Attributes")
     thread_logger.info(f"Edit_User RESPONSE ->>>> {updated_item}")
     return updated_item
-    
-def get_all_users():
-    thread_logger.info("Attempting to retrieve all users.")
-    response = users_table.scan()
-    thread_logger.info(f"Get_All_Users RESPONSE ->>>> {response.get('Items')}")
-    return response.get('Items')
-
-def register_entry(tag_id: str, user_id: Optional[str]):
-    entry = {
-        "LogID": str(generate_unique_id()),
-        "TagID": tag_id,
-        "UserID": user_id,
-        "Time": datetime.datetime.now(datetime.timezone.utc).isoformat()
-    }
-    access_log_table.put_item(Item=entry)
-
-    if user_id:
-        user = get_user(user_id)
-        if user:
-            user['LastScanned'] = datetime.datetime.now(datetime.timezone.utc).isoformat()
-            users_table.put_item(Item=user)
-
-
-def get_entries_count(user_id: str):
-    response = access_log_table.scan(
-        FilterExpression="UserID = :user_id",
-        ExpressionAttributeValues={":user_id": user_id}
-    )
-    return len(response.get('Items', []))
